Fill out Meta struct
Fill out Meta by storing created time, last modified time, and the
version string. Note that version is unimplemented here and should be
ignored by the provisioning client.

This creates a new set of "Stored" objects, starting with StoredMeta,
which are the parts of Meta the provider store will need to record (as
opposed to resource type and location, which will be filled in later).

Next, StoredUser and StoredGroup combine User and Group with the fields
of StoredMeta, and these two objects are the new level of abstraction
the provider store needs to care about.

One additional notable change in this commit is that `ListResponse` is
now implemented as a list of `SingleResourceResponse`s instead of
resources.

Closes #4